+++ code/core/src/phone/java/com/adobe/marketing/mobile/services/ui/notification/AppIconHelpers.kt
@@ -9,7 +9,7 @@
   governing permissions and limitations under the License.
 */
 
-package com.adobe.marketing.mobile.services.ui.pushtemplate
+package com.adobe.marketing.mobile.services.ui.notification
 
 import android.content.Context
 import android.content.pm.PackageManager
