+++ code/core/src/phone/java/com/adobe/marketing/mobile/services/ui/notification/BasicPushTemplate.kt
@@ -9,11 +9,7 @@
   governing permissions and limitations under the License.
 */
 
-package com.adobe.marketing.mobile.services.ui.pushtemplate
+package com.adobe.marketing.mobile.services.ui.notification
 
-/**
- * Exception indicating that construction of a push notification failed.
- *
- * @param message [String] containing the message for the new exception
- */
-class NotificationConstructionFailedException(message: String?) : Exception(message)
+internal class BasicPushTemplate(messageData: MutableMap<String, String>) :
+    AEPPushTemplate(messageData)
