+++ code/core/src/phone/java/com/adobe/marketing/mobile/services/ui/notification/NotificationChannelHelpers.kt
@@ -9,7 +9,7 @@
   governing permissions and limitations under the License.
 */
 
-package com.adobe.marketing.mobile.services.ui.pushtemplate
+package com.adobe.marketing.mobile.services.ui.notification
 
 import android.app.NotificationChannel
 import android.app.NotificationManager
